Separate file to rethrieve and update data - Cleanup and more consistant code
